Update warpstone model to use same type as Location

+++ src/main/java/com/epic/warpstones/models/Warpstone.kt
@@ -3,9 +3,9 @@ package com.epic.warpstones.models
 import java.util.UUID
 
 class Warpstone {
-    public var x: Int = 0
-    public var y: Int = 0
-    public var z: Int = 0
+    public var x: Double = 0.0
+    public var y: Double = 0.0
+    public var z: Double = 0.0
 
     public var name: String = ""
     public var destination: String = ""
